# MAEBACKEND/Endpoints/openapi_specification.yaml
openapi: 3.0.3
info:
  title: EasyRent API - Sistema Completo de Marketplace Inmobiliario
  description: |
    API completa para el marketplace inmobiliario EasyRent con todas las funcionalidades:
    - 🔐 **Autenticación**: JWT + API Keys (Compatible con base de datos PostgreSQL)
    - 👥 **Gestión de usuarios y agencias**
    - 🏠 **Listings y multimedia**
    - 🔍 **Búsqueda avanzada y filtros**
    - 💎 **Sistema de planes y facturación**
    - 📊 **Analytics y reportes**
    - ✅ **Verificación y moderación**
    - 🔗 **Integraciones externas**
    
    ## Tecnologías
    - **Framework**: FastAPI + Python 3.11+
    - **Base de datos**: PostgreSQL 17
    - **Cache**: Redis
    - **Message Broker**: Apache Kafka
    - **Autenticación**: JWT (Bearer Token)
    
    ## Notas de Implementación
    - **Autenticación Firebase**: Sistema de autenticación completo con Firebase Authentication
    - **Flujo de autenticación**: Frontend → Firebase → Backend verification → JWT tokens
    - **Sin manejo de passwords**: Toda la gestión de credenciales delegada a Firebase
    - **Soporte híbrido**: Funciona con o sin Firebase configurado (modo desarrollo)
    - **Schemas actualizados**: Los modelos de datos coinciden con la implementación real en FastAPI
    
    ## Autenticación Firebase
    Para autenticarse, el cliente debe:
    1. Autenticarse con Firebase (frontend)
    2. Obtener el Firebase ID token
    3. Enviar el token al endpoint `/auth/login`
...
    
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /listings/{listing_id}/images:
    get:
      tags: [Media]
      summary: Obtener imágenes de propiedad
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de imágenes

    post:
      tags: [Media]
      summary: Subir imágenes
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                descriptions:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Imágenes subidas
  /listings/{listing_id}/images/{image_id}:
    put:
      tags: [Media]
      summary: Actualizar imagen
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                is_primary:
                  type: boolean
                order_index:
                  type: integer
      responses:
        '200':
          description: Imagen actualizada

    delete:
      tags: [Media]
      summary: Eliminar imagen
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Imagen eliminada
  /listings/{listing_id}/videos:
    get:
      tags: [Media]
      summary: Obtener videos de propiedad
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de videos

    post:
      tags: [Media]
      summary: Subir video
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
                title:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Video subido
  /listings/{listing_id}/videos/{video_id}:
    put:
      tags: [Media]
      summary: Actualizar video
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Video actualizado

    delete:
      tags: [Media]
      summary: Eliminar video
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video eliminado
  /media/upload-url:
    post:
      tags: [Media]
      summary: Obtener URL de subida directa (S3 presigned)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, content_type]
              properties:
                filename:
                  type: string
                content_type:
                  type: string
                size:
                  type: integer
      responses:
        '200':
          description: URL de subida y metadata

  # ===========================================
  # USER INTERACTIONS
  # ===========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación usando JWT Token.
        
        **Formato**: `Authorization: Bearer <token>`
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Autenticación usando API Key para integraciones.
        
        **Formato**: `X-API-Key: <your-api-key>`
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticación usando JWT Token.
        
        **Formato**: `Authorization: Bearer <token>`
        
        **Ejemplo**: `Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...`
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Autenticación usando API Key para integraciones.
        
        **Formato**: `X-API-Key: <your-api-key>`
...